requirements.txt
# Minimal working set for Python 3.13
python-dotenv
fastapi
uvicorn

# Pinecone with latest client
pinecone

# LangChain - using older compatible versions
langchain
langchain-core
langchain-text-splitters
langchain-google-genai
langchain-pinecone

# Utilities
pypdf
requests
python-multipart
langsmith
pydantic

# CLIP Embeddings & Image Processing
sentence-transformers
Pillow
pymupdf

# Cloudflare R2 Storage (S3-compatible)
boto3
